upload distractorTest.js
add a distractor test to test against generateGreenInRed and generateRedInGreen funcions. To run this test, right-click a webpage and select Console tab. In the tab, copy and paste the content from this test and will output the number of times the distractor shows up in every location:  {top: 1988, mid_right: 1974, bot_right: 1974, bot_left: 2044, mid_left: 2020}